No Drag Zone Hazard
Made a hazard that is a zone that when the "hero" is dragged through, will un-tag him as a draggable unit so long as he is in the zone. Should they exit the zone, they will be retagged as draggable. If the hero is stationary in the zone, they will be reset to a location with an explosion particle effect signaling this.